rework structure of gradio app
