+++ scraibe/app/__init__.py
@@ -1,2 +1,7 @@
 from .qtfaststart import *
-from .gradio_app import *
+from .activity_tracker import *
+from .interface import *
+from .stg import *
+from .interactions import *
+from .global_var import *
+from .app import *
