+++ scraibe/app/global_var.py
@@ -1,0 +1,9 @@
+"""
+Stores global variables for the app.
+"""
+
+# Global variable to store the model
+MODEL = None
+
+# Global variable to track user activity
+USER_ACTIVE = False
